First pass cleaning up Polygon.
Removed THREADED option with mutexes for computing edge normals on the
fly. This likely wasn't a good optimization in the first place for
multithreaded programs, and contained a bug to boot.

Simplified some of the interface and added a few constraints.

Some of the interface is still painfully elaborate. Needs more work.